Renamed and re-referenced .h and .c
Renamed corresponding .h and .c files in each subfolder to match the folder name (ortho/staggered). also, changed the referenced file in each .c file correspondingly.
