+++ keyboards/sneakbox/disarray/ortho/ortho.c
@@ -1,5 +1,5 @@
 /*
-Copyright 2021 Bryan Ong
+Copyright 2020 Bryan Ong
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -15,5 +15,4 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "disarray.h"
-
+#include "ortho.h"
